tp_docker/backend/src/requirements.txt
pandas == 2.2.0
numpy == 1.25.2
mysql-connector-python == 8.4.0 